Reduce the number of times we run apply
The current implementation of the install scripts runs `terraform apply`
2 times per terraform module using -target. Additionally apply is run
  again 2 more times at the end of the script.

This means that we are running terraform apply twice as much as
required.

This commit adds ` | tee /dev/tty` to all the apply steps so that the
output can be captured and evaluated, as well as having the output print
to screen.